refactor : strikeCount ballCount 메서드 checker 클래스로 분리
