+++ src/main/java/gmbs/Baseball.java
@@ -7,26 +7,7 @@ public class Baseball {
     private final UserInput userInput = new UserInput();
     private final RandomNumberGenerator random = new RandomNumberGenerator();
     private final Display display = new Display();
-
-    public int strikeCount(List<Integer> randomNumbers, List<Integer> userInput) {
-        int strikeCount = 0;
-        for (int i = 0; i < userInput.size(); i++) {
-            if (userInput.get(i).equals(randomNumbers.get(i))) {
-                strikeCount++;
-            }
-        }
-        return strikeCount;
-    }
-
-    public int ballCount(List<Integer> randomNumbers, List<Integer> userInput) {
-        int ballCount = 0;
-        for (int i = 0; i < userInput.size(); i++) {
-            if (randomNumbers.contains(userInput.get(i)) && !userInput.get(i).equals(randomNumbers.get(i))) {
-                ballCount++;
-            }
-        }
-        return ballCount;
-    }
+    private final BaseballChecker checker = new BaseballChecker();
 
     private boolean isStrikeOut(int strikeCount) {
         return strikeCount == Constant.INPUT_LENGTH;
@@ -38,8 +19,8 @@ public class Baseball {
         List<Integer> randomNumbers = new ArrayList<>(random.getRandomNumbers(Constant.INPUT_LENGTH));
         do {
             List<Integer> inputNumbers = new ArrayList<>(userInput.getNumbers());
-            strikes = strikeCount(randomNumbers, inputNumbers);
-            balls = ballCount(randomNumbers, inputNumbers);
+            strikes = checker.strikeCount(randomNumbers, inputNumbers);
+            balls = checker.ballCount(randomNumbers, inputNumbers);
             display.showStrikeBallCounts(strikes, balls);
         } while (!isStrikeOut(strikes));
     }
